Update README.md
•	Through student teacher platform we can digitalize the present education system. It will help in letter student gain better through video tutorials and it will minimize the use of papers and will be effective in saving trees. Making the leaning process accessible anywhere to the student. The achievement of the above objectives can help in managing the resources efficiently. The automated process will lead to time saving and eradication of common errors.

•	 To increase access to learning opportunities/increase flexibility for students
•	To enhance the general quality of teaching/learning.
•	To develop the skills and competencies needed in the 21st century, and in particular to ensure that learners have the digital literacy skills required in their discipline, profession or career – or, put simply, to get work in the future
•	To meet the learning styles/needs of millenial students
  

CONCLUSION:
Implementing the Online Student Course Registration System, the registration procedure has been simplified. The students can access the registration portal online either from a computer or a smart phone, and get the necessary information of any subject. This web application provides us with ease of access, user friendliness and transparency. On the other hand, from organizations viewpoint, it helps in maintaining transparency, data consistency, data accessibility and easy maintenance. Implementation of the proposed system will reduce the workload of all those involved as the data can be now managed with proper authentication and authorizations instead of being hard copied and accessible to everyone. All the technologies i.e. PHP, JavaScript and MySQL used for current system design are open source and hence freely available for download. PHP provides a strong platform for creating the visual front-end of the web application and PHP combined with HTML provides a very flexible development environment. For the purpose of fulfilling Web Server requirements WAMPP was used, which is again Open Source and is supported across multiple platforms. The primary objective of Student Teacher platform was to automate student learning procedure. It has been achieved successfully and the system is tested to be working efficiently.